Reorganized project to fit the package layout.
