+++ app/src/main/res/layout/activity_product_screen.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ProductScreenActivity">
+    tools:context=".ui.activities.ProductScreenActivity">
 
     <TextView
         android:id="@+id/itemNameTextView"
