+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -7,12 +7,12 @@
 
     <fragment
         android:id="@+id/nav_home"
-        android:name="com.theroughstallions.genki.ui.home.HomeFragment"
+        android:name="com.theroughstallions.genki.ui.fragments.HomeFragment"
         android:label="@string/menu_shopping_list"
         tools:layout="@layout/fragment_home" />
     <fragment
         android:id="@+id/nav_settings"
-        android:name="com.theroughstallions.genki.ui.settings.SettingsFragment"
+        android:name="com.theroughstallions.genki.ui.fragments.SettingsFragment"
         android:label="@string/menu_settings" />
 
 </navigation>
